marshal: soft reference for cached serialized bytearray
Under memory pressure, the marshal should drop its cached serialized
view of its self. This is already fairly unfortunate as the footprints
of marshals in memory is at least 2x the inputs and creates memory
churn. Performance is less predictable under pressure, but at least
it'll continue to work.

This optimization is worth the cost as it drastically reduces the cost
of the compareTo operation as used in the Apache Crunch in memory
pipeline.